refined model generation
