+++ sdk/agentserver/azure-ai-responses-server/azure/ai/responses/server/_validation.py
@@ -2,10 +2,10 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from typing import Any, Mapping
 
 from ._options import ResponsesServerOptions
+from .models.errors import RequestValidationError
 
 try:
     from ._generated import ApiErrorResponse, CreateResponse, Error
@@ -24,35 +24,6 @@ try:
     from ._generated import _validators as _generated_validators
 except Exception:  # pragma: no cover - optional until validator generation is integrated in all environments.
     _generated_validators = None
-
-
-@dataclass(slots=True)
-class RequestValidationError(ValueError):
-    """Represents a client-visible request validation failure."""
-
-    message: str
-    code: str = "invalid_request"
-    param: str | None = None
-    error_type: str = "invalid_request_error"
-    debug_info: dict[str, Any] | None = None
-
-    def __post_init__(self) -> None:
-        """Initialize the parent :class:`ValueError` message."""
-        ValueError.__init__(self, self.message)
-
-    def to_error(self) -> Error:
-        """Convert this validation error to the generated ``Error`` model."""
-        return Error(
-            code=self.code,
-            message=self.message,
-            param=self.param,
-            type=self.error_type,
-            debug_info=self.debug_info,
-        )
-
-    def to_api_error_response(self) -> ApiErrorResponse:
-        """Convert this validation error to the generated API error envelope."""
-        return ApiErrorResponse(error=self.to_error())
 
 
 def parse_create_response(payload: Mapping[str, Any]) -> CreateResponse:
