+++ sdk/agentserver/azure-ai-responses-server/azure/ai/responses/server/models/_generated/_models.py
