+++ sdk/agentserver/azure-ai-responses-server/azure/ai/responses/server/models/_generated/sdk/models/models/_enums.py
